+++ test/test_hypothesis/test_connection.py
@@ -2,11 +2,14 @@ import pytest
 
 from .base import BaseMachine, commands, dbnums, run_machine, st, values
 
-connection_commands = commands(st.just("echo"), values) | commands(st.just("ping"), st.lists(values, max_size=2))
+connection_commands = (
+    commands(st.just("echo"), values)
+    | commands(st.just("ping"), st.lists(values, max_size=2))
+    | commands(st.just("select"), dbnums)
+)
 
 
 class ConnectionMachine(BaseMachine):
-    # TODO: tests for select
     base_commands = connection_commands
     redis_only_commands = commands(st.just("swapdb"), dbnums, dbnums)